Add param_feature support
